Refactor: PointMaterial.bsdf
