+++ result.txt
@@ -2,6 +2,6 @@ num threads: 8.
 resolution: 960x540.
 max sampling: 20x4 spp.
 time limit: 14400.00 sec.
-init scene: 0.33 sec.
+init scene: 0.34 sec.
 sampled: 20x4 spp.
-total 46.492000000000004 sec. used 0.32 % (x 309.73)
+total 57.657000000000004 sec. used 0.40 % (x 249.75)
